backend/app/requirements.txt
# App
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.3.4
di4injector==0.0.2      # DI
slf4py==0.0.4           # Logger
html2text==2024.2.26    # Mail
requests==2.31.0        # HTTP Request
python-dotenv==1.0.1    # Environment
pytz==2024.1            # Timezone
cachetools==5.3.2       # Cache
SQLAlchemy==2.0.35      # DataBase
mysqlclient==2.2.4      # MySQL
psycopg2-binary==2.9.9  # Postgresql
psycopg2==2.9.9         # Postgresql
redis[hiredis]          # Redis

# Common
google-cloud-pubsub==2.26.1
google-cloud-storage==2.18.2

# API Gateway
PyJWT==2.9.0  # JWT

# Authority
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Market
pandas==2.2.3           # Data Analytics
boto3==1.35.37          # AWS
fake_useragent==1.4.0   # Scraping
beautifulsoup4==4.12.3  # Scraping